apps/api/pyproject.toml
[project]
name = "ayekta-api"
version = "0.1.0"
description = "Ayekta v2 sync + future clinical API — Postgres (Supabase / RDS / Cloud SQL compatible)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.109.0",
  "uvicorn[standard]>=0.27.0",
  "asyncpg>=0.29.0",
  "pydantic-settings>=2.1.0",
  "pyjwt[crypto]>=2.8.0",
  "bcrypt>=4.1.0",
]

[project.optional-dependencies]
dev = ["httpx>=0.26.0", "ruff>=0.2.0"]
storage-s3 = ["boto3>=1.34.0"]

[build-system]
requires = ["hatchling>=1.13.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ayekta_api"]
